Clean-up redux logic for NoteDetails and add ESDoc
