+++ src/redux/index.ts
@@ -15,7 +15,7 @@ import { DarkModeConfigReducer } from './dark-mode/reducers'
 import { DarkModeConfig } from './dark-mode/types'
 import { EditorConfigReducer } from './editor/reducers'
 import { EditorConfig } from './editor/types'
-import { NoteDetailsReducer } from './note-details/reducers'
+import { NoteDetailsReducer } from './note-details/reducer'
 import { NoteDetails } from './note-details/types'
 import { UserReducer } from './user/reducers'
 import { OptionalUserState } from './user/types'
